LayoutTests/editing/selection/ios/tap-to-set-selection-in-editable-web-view-expected.txt

This test verifies that the selection changes when initially focusing an editable web view by tapping, if the WebKit client unconditionally allows programmatic focus to show the keyboard. To manually test, tap the red square above and verify that the selection is set

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS clickedTarget became true
PASS getSelection().rangeCount is 1
PASS getSelection().type is 'Caret'
PASS target.contains(selectionRange.startContainer) is true
PASS target.contains(selectionRange.endContainer) is true
PASS successfullyParsed is true

TEST COMPLETE
